requirements.txt
pytest==8.3.4
google-auth==2.18.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
pytz==2022.7.1
slack_sdk==3.27.1
Flask==3.0.3
langchain-core==0.3.28
pydantic==2.10.3
langchain-ollama==0.2.2
langchain==0.3.13
pytest-cov==6.0.0
python-dotenv==1.0.1